updated urls and views.py of each app
