+++ ProjetPython/urls.py
@@ -15,20 +15,16 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path
+from django.urls import path, include
 
 from client import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', views.home, name='home'),
-    path('login/', views.login_view, name='login'),
-    path('signup/', views.signup, name='signup'),
-    path('logout/', views.logout_view, name='logout'),
-    path('dashboard/', views.dashboard, name='dashboard'),
-    path('settings/', views.settings, name='settings'),
-    path('client/',views.clientpage,name='clientPage'),
+    path('client/', include(('client.urls', 'client'), namespace='client')),
+    path('freelancer/', include(('freelancer.urls', 'freelancer'), namespace='freelancer')),
 
+    path('', views.home, name='home'),
 
 
 ]
