+++ client/templates/client/profile.html
@@ -5,8 +5,5 @@
     <title>Title</title>
 </head>
 <body>
-
-    
-
 </body>
 </html>
