+++ freelancer/urls.py
@@ -1,0 +1,13 @@
+from django.urls import path
+from . import views
+
+
+urlpatterns = [
+
+    path('dashboard/', views.dashboard, name='dashboard'),
+    path('settings/', views.settings, name='settings'),
+
+
+
+
+    ]
