+++ templates/navbar.html
@@ -15,7 +15,7 @@
                 <div class=" d-flex justify-content-center" style="margin-right: 15rem;margin-left: 25rem">
                     <ul class="navbar-nav">
                         <li class="nav-item">
-                            <a class="nav-link fw-bold" href="{% url 'dashboard' %}">Explore</a>
+                            <a class="nav-link fw-bold" href="{% url 'freelancer:dashboard' %}">Explore</a>
                         </li>
 
                         <li class="nav-item dropdown">
@@ -46,10 +46,10 @@
                     {% if not logged_in %}
                         <ul class="navbar-nav">
                             <li class="nav-item me-3">
-                                <a class="nav-link" href="{% url 'login' %}">Login</a>
+                                <a class="nav-link" href="{% url 'client:login' %}">Login</a>
                             </li>
                             <li class="nav-item">
-                                <a href="{% url 'signup' %}">
+                                <a href="{% url 'client:signup' %}">
                                     <button type="button" class="btn btn-primary">Sign up</button>
                                 </a>
                             </li>
@@ -71,7 +71,7 @@
                             </a>
                             <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="profileDropdown">
                                 <li><a class="dropdown-item" href="#">View Profile</a></li>
-                                <li><a class="dropdown-item" href="{% url 'settings' %}">Settings</a></li>
+                                <li><a class="dropdown-item" href="{% url 'freelancer:settings' %}">Settings</a></li>
                                 <li>
                                     <hr class="dropdown-divider">
                                 </li>
